+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Feb 27, 2012
+
+* Refactor request processing - add more tests.
+
 ## Feb 26, 2012
 
 * Refactor request processing - this work is in progress.